email_aws_ses/requirements.txt
amqp==5.3.1
billiard==4.2.1
blinker==1.9.0
boto3==1.38.23
botocore==1.38.23
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
DateTime==5.5
dnspython==1.16.0
Flask==3.1.1
flask-cors==6.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.3
MarkupSafe==3.0.2
pillow==11.2.1
postmarker==1.0
prompt_toolkit==3.0.51
pymongo==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.1.0
requests==2.32.3
s3transfer==0.13.0
setuptools==80.8.0
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
zope.interface==7.2
APScheduler==3.11.0
gunicorn==23.0.0
